Add MovementModel

+++ src/main/java/net/pkhapps/dispatchsimulator/engine/geo/Distance.java
@@ -6,5 +6,5 @@ package net.pkhapps.dispatchsimulator.engine.geo;
  * @param distanceInMeters the distance in meters.
  * @see Speed
  */
-public record Distance(int distanceInMeters) {
+public record Distance(long distanceInMeters) {
 }
